test(cli): exercise managed account drift through native mutation fixture
